Changes the method of acquiring a track. Fixes a crash.
If the track is to be skipped or the MP3 is cached, the track
information won't be loaded.

Fixes crash when there's no MP3 url in track page. An album might list
a track that has no MP3 url on its page. This is not possible to
determine until the track page is loaded. After loading it, if there's
no MP3, we skip the track.